Updated some behaviour for the base provider
Make sure we're using new_resource where we need to be.

I've also realised that as l_c_r isn't called before every action (a
potentially intensive method) there is no way to determine wether a
service is running or not. simply setting @current_resource.running when
start a service is a little silly, too.. soo I'll have to look into
this.