Restructure the package provider to use the candidate_version accessor vs the variable directly.
This allows us to delay loading the candidate_version until it's actually needed.

For example, it's not needed if we didn't specify a version, we're doing an install, and we already have a version installed.

If we do this, then we can possibly speed up the other providers because they don't have to look up a candidate version, only
an actual installed version. Some can do this via disk instead of via the package maintainers, and this makes it go faster.