Attrs for flexible callbacks in deploy resource.
Attributes for callbacks:
* :before_migrate
* :before_symlink
* :before_restart
* :after_restart

Each attribute can take a block or hash. Block is to be executed
as an inline recipe. Hash specifies a file to be read as an inline
recipe or instance_eval'd in the context of the provider.
instance_eval option is the default, preserves backwards compat
with chef-deploy and capistrano-ported deployments.