Make setting the run_list override it, not append to it
Previously, the run list was handled specially and you could only
append to it from a json file.  Make it so that setting it in the json
file now overrides the list on the server instead.

Also, clean up the node a little bit by making recipes an alias for
run_list and getting rid of a now-unused method.