Fixes CHEF-926 - deploy_revision resource no longer deletes current release
The deploy revision resource previously always appended
the current release to the release cache list. As
a result, if the current release were deployed
enough times it could eventually occupy a slot
outside of the release that are kept during
the call to cleanup! and be deleted. This fix
ensures that the release is removed from the
release cach list before being appended again to
prevent it from being deleted.