[CHEF-1514] don't accidentally flatten the users data and then Hash() it
value_for_platform was creating an intermediate representation of the
platform specific values as an array, and then flattening it to make it
suitable for creating a hash using the Hash[] method. This was
flattening the user data which caused errors when that data is an array.
Fix it by only flattening one level.