[CHEF-1485] workaround SSH hangs when bootstrapping on EC2
modify ec2 server create so that it waits for sshd to come up on the
host, *THEN* waits a bit for it to generate its host key, then continues
on to bootstrapping. This ensures that we're sleeping during the host
key generation process no matter how long the instance takes to boot.