Remove extlib dependency, copy mash code into sources, fix CHEF-1405
Including extlib in Chef introduces conflicts for those wanting to
access Chef via Rails applications.  Only the Mash class is used by Chef
and this functionality has been extracted from the extlib gem and added
to the Chef sources.

Thanks to Wojciech Wntrzak and Aaron Suggs for proposing and testing
this solution.