Move knife cloud commands to plugins
Knife cloud commands now live in their own repos:
* https://github.com/opscode/knife-ec2
* https://github.com/opscode/knife-bluebox
* https://github.com/opscode/knife-slicehost
* https://github.com/opscode/knife-rackspace
* https://github.com/opscode/knife-terremark

This gives us the ability to specify fog as an actual dependency for
these projects, so we can expect a consistent version. Also, users will
only need to install the commands for the clouds they use, which will
make the output of `knife --help` a little more manageable.