Keep support for '*' at the start of a search query
A query like 'ec2:*' is not valid lucene query syntax, however,
because of the format of our search index and the way we transform
queries, we can support it.  It makes sense to enable this as these
queries are convenient and seem to match user's expectations.

Example queries are now tested in sorted order