Don't call 'gem update --system' in bootstrap
Updating Rubygems can cause problems if the new version of Rubygems
has not yet been tested.  The bootstrap script installs Rubygems 1.6.2
which is known to work well with Chef and its dependencies.  If a
newer/latest version of Rubygems is desired, this can be achieved via
a cookbook early in the runlist.