[CHEF-2947] properly remove STDERR from list of fds to select/read
Properly remove child's stderr pipe from the list of IO objects to
select on in the case that stdout has already been removed from the
list. This prevents a situation where the stdout stream is always
considered readable by select (because it has reached eof) but ShellOut
would attempt to read from it in a tight loop. Since GC is disabled
while a command is running, this would generate a bunch of EOFError
objects that were not reaped and use a ton of heap space.