Multicookbook upload support
When I originally wrote this a few months ago, for 29 cookbooks, the majority of
which only have a single recipe (i.e. the speedup would be minimal) we move from
38 seconds to 17 seconds.

Now I have 39 cookbooks which have more meat and the time goes from 42 seconds
to 14 seconds.

Note that it still does the more naive approach for when you list cookbooks.
This is becuse the better error reporting is worthwhile when uploading
hand-picked cookbooks even if there are 2 or 3, but when re-uploading everything
then the batched approach is a huge improvement.