[CHEF-3143] - Omnibus path and gem binary fix
The check for the omnibus path should be a regular expression so it
can detect either /opt/opscode or /opt/chef locations in the event
that a user upgraded an existing chef-full installation by installing
the gem.

The @new_resource.gem_binary path should be set to the detected gem
binary location (via find_gem_by_path), else it will be nil and cause
the gem to be installed through the API, instead of using the binary.