Fix broken whyrun unit test for user resource
This initially appeared as a failing test user Ruby 1.8 with an error of
"undefined method `sp_pwdp' for nil:NilClass." This was ultimately because
we were doing a lookup for user info for 'adam' which didn't exist. However,
getpwsnam will return nil if you're not root and much time was spent reworking
the pending blocks which are particularly tricky and usually are not doing
what you think they're doing.

Running these tests under bundler will require the ruby-shadow gem, so I
have added that as a development dependency.