Refactor remote directory provider for Windows
Windows treats a symlinked directory as a directory
whereas linux treats the symlink as a file. Thus on windows
we have to remove the directory as if it were a directory.

Also, we can only create links on windows if we are an Administrator,
so I've added a pending rescue to one test accordingly.