CHEF-1554: default to using a cloud attribute
If we have a cloud attribute we should use that, but the user should
still be able to override that on the command line. This is a little tricky
because we use the same variable for options passed from the user as we do
to represent our final choice. Thus we lose any history as it whether our
final choice was user requested or a default.

Alternately we could refactor the class to use a different variable as a
result of #configure_attribute