CHEF-2496: Move accept_license attribute to ips_package resource
The accept_license attribute is unique to IPS, at least in function. Thus,
I have moved it over to the ips_package resource. This is a unique enough
edge case that nobody knows of a package that --accept is needed for, so
mostly accept_license is about avoiding a situation where we accept a license
for the user without their knowledge.