[CHEF-581] Always run delayed notifications
Delayed notifications will now be run even if the "body" of the chef run
fails, or if a failure occurs in another delayed notification.

The justification is that delayed notifications are commonly used to
make changes to state that is opaque to chef so a subsequent run will
not detect that the resource's state was never corrected. For example,
chef has no way to determine whether the in-memory state of a service is
correct, so if chef drops a delayed restart notification for that
service, running chef again will not bring the service into the correct
state.