CHEF-3435: Create deploy dirs before calling scm_provider
The SCM providers have an assertation that requires the deploy directory to
exist. The deploy provider will create missing directories, we don't converge
the actions before we call run_action against the SCM provider, so it is not
yet created. This ensures we run any converge actions waiting before we call
the SCM provider.