[CHEF-2936] add role and environment components to Node::Attribute
I reformatted Chef::Node::Attribute because the indentation was busted.
The actual code change is that Attribute now has components for
role_default, env_default, role_override, and env_override, and merges
them according to the precedence given in the documentation.

These new components are not yet wired in to Chef::Node, however.