[CHEF-2936] ensure all default and override attrs go in node json
default and override attributes are now split into components, so all
the components need to be merged together to generate the serialized
representation of a node. Unfortunately, this means that serializing a
node is now lossy because the information about which component default
and override attributes belong to is lost. In practice, this is not a
major issue, since automatic, override, and default attributes are
cleared by chef for each run.