CHEF-3577: Allow ssh config port and user to be used
Net::SSH::Multi takes a user argument which we pass, so we need to read
the user from the ssh_config before passing the user and use it if
applicable.

We also were sending port 22 even if the user hadn't specified a port,
due to a mixlib-cli default. This would override any port set in an
ssh config. It isn't necessary to send a port, Net::SSH will default to
22 on its own.

Credits: @btm