Switch template provider to FileUtils.cp
This reverses the changes made in CHEF-1396. When copying a file over
an existing file, the permissions of the existing file are maintained.
When moving the file the permissions of the source file are preserved.
Since CHEF-1396, we now set permissions after copying (formerly moving)
the file, but only do so if there are permissions specified by the user
to be set.