Route system status JSON endpoint to /_status instead of /status
The webui has a /status path already. Using _status avoids this
conflict and future conflicts with the API. The leading underscore
marks this as "special" which is appropriate for system status.

Doing this allows us to continue using the X-Ops-UserId header as the
switching mechanism between webui and api and expose system status in
a way that is trivial to retrieve using curl or a browser (where
otherwise, you'd have to set the x-ops-userid header).