Quote cookbook name/version in error message
User was getting an error message:
> ERROR: Cookbook git depends on cookbook runit, version >= 0.0.0,
Issue was the metadata for git had a dependency on 'runit,'. Should
probably also add validation to metadata depends for cookbook names, but
generally quoting user-supplied values in error messages makes this kind
of issue more obvious, so this should be a worthwhile quick fix.