[CHEF-3662] ApiClient can set a private key from JSON
Without this change it's not possible to fix CHEF-3662 on the client
without modifying Chef::REST to ignore json_class. Previously the
behavior was explicitly defined to be the opposite, but the reasoning
behind this decision wasn't documented; I suspect it may have been for
server-side behavior that's no longer necessary.