Add Chef::User class
This commit adds a Chef::User class similar to the Chef::ApiClient
class.  Notable differences are the presence of a password attribute
and the fact that the API will not return chef_type or json_type
attributes.  The latter means that the Chef::REST class will not
automatically inflate API responses to Chef::User objects.