Peg software components at current SHAs
While tedious, our current build pipeline will otherwise emit and
publish package names with content that may differ in terms of
component versions in important ways (with/without a bug fix). By
forcing at least git SHAs, we'll force a commit in this repo, and thus
a new build number.

Aside from a wanting a lock file for components, it would also seem
reasonable to put a portion of a digest of the version manifest into
the package name since that is really what is being packaged, not
software definitions in this repo.

Finally, the chef software project is doing something fancy and it
isn't clear how to peg it to a sha.