only set default value for CHEF_GIT_REV if value is unset
This will ensure we don't inadvertently set a value for CHEF_GIT_REV in 
other projects in this repository (i.e. chef-client)! Would be 
preferable to change the 'load the world' project loading logic in 
Omnibus: 
https://github.com/opscode/omnibus-chef/blob/master/Rakefile#L16

We really should only load the project definition for the project we 
want to build.