build webui url off configured values for listen + port
The default attribute value for
`node['chef_server']['chef-server-webui']['url']` is out of date the
instant a user configures alternate values for `listen` or `port`. We'll
remove this misleading attribute and just compute a url when we need it
using the following format:

  http://WEBUI_LISTEN:WEBUI_PORT

`node['chef_server']['chef-server-webui']['listen']` has also been
updated to match the idioms of other components `listen` attribute.