Fixes CHEF-3307
Use metadata name when loading cookbook

If a name attribute is loaded from a cookbooks metadata, use this for
dependency checks and references as specified in documentation. Ensure real
pathname is retained so that manifests can still be generated correctly.

In addition we must read the metadata implicitly when cookbooks are loaded to
ensure we are using the correct name.

Also include a deprecation warning when encountering missing names in metadata
as this is a planned breaking change in Chef 12.