Add a project to build chef in separate directory
We cannot install vanilla omnibus packages on the machines that build or
test omnibus packages because they conflict over the /opt/chef
directory. We could install w/ rubygems but on some platforms omnibus is
the only viable method we have for building chef and all of it's deps.
Therefore, we need a secondary chef package that we can use to manage
the omnibus build and test machines.