pin windows gems to known good versions
This is a quick fix for an issue where unpinned versions of gems are
getting installed on windows. The test phase of the omnibus builds is
using bundler, so the tests use version constraints in the Gemfile,
while chef/knife in the actual package are not running with bundler and
so use the unconstained gem versions as installed by omnibus. A recent
release of the ffi gem had a bug in its error handling, which is causing
problems for chef users on windows. See: CHEF-4145