restore missing assertion in remote_file http spec
Cleans up the test to use let bindings everywhere. Motivation for
changing this file is to fix CHEF-3140 regression test. Test needs to
assert that Chef::REST is getting created with the disable_gzip option.
Because of overlapping method expectations, this assertion was commented
out. To make the tests more resilient to this kind of mistake, expected
arguments were added to the method call expectation for
Chef::REST.new().