CHEF-4197: ensure capture groups are populated.
http://tickets.opscode.com/browse/CHEF-4197

Wrap the alternation groups in an outer group to ensure that the named captures are populated regardless of whether the first or second clause matches. If the device file is a symlink (Ubuntu 12), then the mount options will be incorrectly populated. This may well manifest when moving from Ubuntu 10.x to Ubuntu 12.x.