don't use client time for If-Modified-Since
RFC 2616 notes recommend caution when using client-side time for
If-Modified-Since due to potential time skew and other issues. The RFC
further suggests that servers that do not include Date headers in
responses may not have reliable timekeeping. This suggests that we
should not try to use If-Modified-Since when accessing resources on such
servers rather than falling back to local time.

In addition to addressing mtime issue above, add tests for all header
handling in the remote file HTTP implementation.