Enable conditional GET by default
The upside of this is that users who have not specified a checksum on
their remote file resources will automatically get the benefit of
conditional GET to reduce file fetching as long as the origin server
supports it.

The risk is that some origin servers may have incorrectly implemented
HTTP and do the wrong thing with the cache control headers. We suspect
this should only be the case for a very small minority of users, who
will be able to opt out of the new behavior.