Fix compiling the same cookbook multiple times
cookbook_order generates the order based on the cookbook as a symbol.
This assumption comes from Chef::Recipe.parse_recipe_name which
to_sym's the cookbook name.

This is in contrast to #each_cookbook_dep which yields the cookbook
dependencies as Strings. The symptom of this mismatch is the
ordered_cookbook Array containing the same cookbook twice, once as a
String and once as a Symbol.

This corrects the issue by forcing everything going into
ordered_cookbooks to be a Symbol.