Ignore exit status 12 when deleting test user account
Various versions of `useradd` return a variety of errors with exit code
12 when removing a user with no home directory and given options
indicating the home directory should be removed, regardless of the `-f`
flag. This is fixed in newer versions (such as the version on Ubuntu
13.04, where these tests were originally written), but probably won't be
fixed for older OSes. Accepting exit code 12 should fix the tests on
these platforms.