Use blocking waitpid only on ENOENT
This is a compromise that prevents us from leaving zombie processes
around after a failed exec raises ENOENT, but prevents us from blocking
(possibly forever) on timed-out commands. The latter case can cause
zombies, so this is not optimal for all cases. See MIXLIB-16 for a
proposed enhancement to solve that case.