CHEF-4587 Windows service provider should treat all START_TYPEs != DISABLED as enabled
Currently, the Windows service provider only treats "Automatic (Delayed Start)"
and "Automatic" START_TYPEs as an enabled service. It should also treat
"Manual" as an enabled service.

A "Manual" service is not a disabled service.  A "Manual" service can be started
by another service that depends on it or by an application or by human
intervention.

If we want Chef to ensure a service is truly disabled then the service's
START_TYPE should be set to "Disabled".