Fix disable gzip for Ruby 2.0
If we've not set an accept encoding by the time we're making a request,
set the accept encoding to request "identity" only. Ruby 2.0 seems to
have some sort of built-in gzip support that we need to explicitly
disable.

Note that prior to this patch, we were _always_ advertising gzip
support, but we would skip decompression if disable_gzip had been set,
which only did the right thing accidentally.