Patch webrick timeout handler to avoid test failures
Webick uses a centralized/synchronized timeout manager. It works by
starting a thread to check for timeouts on an interval. The timeout
checker thread cannot be stopped or canceled in any easy way, and it
makes calls to Time.new, which fail when rspec is in the process of
creating a method stub for that method. Since our tests don't rely on
any timeout behavior enforced by webrick, disable the timeout manager
via a monkey patch.