Check before creating a new system mutex on windows.
This is needed when multiple user accounts are running chef-client on systems. This scenario is imminent when they configure chef to run as service on windows which by default runs as "Local System" and when they try to run chef client manually as Administrator.

NOTE: Non-admin users do not wait on the run_lock but can create a run_lock if there is not one already.