Patch ruby 2.0 Net::HTTP proxy code for IPv6
Ruby 2.0 changes the way proxying is implemented in Net::HTTP such that
each request will check the URI for "proxy-ness." This check will
generate a URL from the given address and port without wrapping IPv6
addresses in brackets, which triggers an error in the URI library.

Check for this case on startup and monkey patch the offending method if
necessary.