Ensure Host header is correctly formed
Net::HTTP does not set the Host header correctly when the host is given
as an IPv6 address. Previous testing determined that Chef was setting
the correct Host header in the authentication code, but did not set the
Host header for non-authenticated requests, so the Net::HTTP default
behavior (bug) would be exposed.

Set Host header "manually" for all requests to hide the Net::HTTP bug.