Lock `sdoc` down to `~> 0.3.0`
TL;DR - `sdoc ~> 0.4.0` depends on `rdoc ~> 4.0` and this version of 
`rdoc` breaks our Omnibus builds on Solaris!

The full explanation is going to sound long-winded...just bear with me. 
`rdoc` 0.4.0+ ships with fixture data from MarkdownTest, a Markdown 
validation test suite. Unfortunately this fixture data also has spaces 
in all of the file names. `pkgmk`, the utility we use to generate Chef 
Omnibus packages for Solaris, has a known issues processing file names 
with spaces:

http://compgroups.net/comp.unix.solaris/pkgmk-fails-when-filenames-have-space-chara/295864

Yes, this fix is a metaphor for the age old battle of hipster vs 
greybeard.