CHEF-5018:  specific test for CHEF-5018
This tests actually goes red with the bug and green when its fixed.

It actually worked fine when @contants was an IO object instead of an
Array because the #each method worked identially on both.  However,
there's a monkeypatch in the wilds that did @contents << "thing" and
that broke when it became an IO object accidentally.