chef_gem: should use omnibus installed `gem`.
In the event that string options are passed to chef_gem then the logic
states that we will shell out and invoke the gem binary.  The problem
with this is it picks up whatever gem is installed in the PATH which is
probably not the omnibus version...

This fix simply uses the bundled `gem` binary when using chef_gem which
is the desired behavior.