Retire the GC disable hack for all but ruby 1.8.7
Originally we needed this hack for 1.8.6 and 1.8.7, which were the most
common versions in use at the time. 1.8.6 is dead so we don't need to
account for it any more, which leaves 1.8.7 as the only case we need to
care about for this hack. On Ruby 1.9 and greater this hack isn't needed
and causes memory bloat for very long running shell commands.