CHEF-4888: Call WIN32OLE.ole_initialize in sub-threads
An exception is raised when loading WIN32OLE (including via ruby-wmi) in one
thread and used in a sub-thread without calling ole_initialize in that thread.
This is common in irb/chef-shell's sub-shells that are created in recipe_mode
and attribute_mode. Since it will be a while before this is fixed in Ruby
and ruby-wmi/rdp-ruby-wmi are stale, we workaround the bug here.

Ruby issue: https://bugs.ruby-lang.org/issues/2618