add sleep to install-test for windows
after releasing artifacts to preprod we need to sleep to wait for
the omnitruck cronjob to scrape s3 and offer the artifacts.  otherwise
the windows slaves always fail the first time through the release
pipeline.

using powershell because we seem to have that everywhere now.  this
whole stupid bat file now should be rewritten in powershell...