CHEF-5147: Fix tests, but...
This fixes the tests for the behavior specified in CHEF-5147 that makes it so
that --include-dependencies will upload dependencies if they're on disk, but
and fail if they're not available on disk or the server.

However, overall this change means that 'knife cookbook upload' no longer
makes a dependency check, which was something we intentionally added so you
would know about dependency problems at cookbook upload time rather than
later when you run chef-client.

See the ticket for further discussion