CHEF-5100: moar func tests
This tests the reuse of the content length validation object and
the situation where the content length of a previous download is
retained while a 403 or other HTTP error fails to re-init the
content length to zero.  The result is the 403 is turned into an
inaccurate content length exception, masking the real error.