CHEF-5247: Add a unit test
Unfortunately the way that Chef::Provider::User::Solaris subclasses
Chef::Provider::User::Useradd we call both Solaris#manage_password
and Useradd#manage_user whenever we call Solaris#manage_user. If we
are updating a password, we only want Solaris#manage_password, so we
have a kind of convulted way set to tp turn Useradd#manage_user into
a noop. This test tries to ensure we can noop in the way we want to,
but we really need a functional or integration test here.