do not accept exitstatus == 1 as success
status.exitstatus of 1 is a failure. In fact, it's a very common exit code for failures on cron.  Unless there is a specific reason that exit status 1 is supposed to be considered okay — and I strongly suspect that there is none — then it should not be treated as okay.  I think this is just an off-by-one error and that >= was meant instead of >, or 0 instead of 1.

We can avoid the question by using Process::Status#success? instead.