Add url_for_checksum helper function to CookbookVersion class
This function makes it easier for 3rd-party users of this class to
determine the url for a given checksums.  Since the URL is not
standard across server types (ChefZero uses file_store, ChefServer
uses bookshelf), the only way to get his information reliably is to
read it out of the metadata.