Print deprecation notices for chef-repo rake tasks
These rake tasks get loaded by the opscode/chef-repo's Rakefile. I
propose that be removed, as the functionality of these rake tasks is
superseded by other tools as noted in the output displayed now for
each task. The functionality is either built into knife, is provided
by a cookbook, or is provided by other tools in the Chef ecosystem,
all of which are included in the ChefDK.

This supersedes #1619.